AI-service/requirements.txt
# ========== AI-service 통합 의존성 (ml-server, recommendation, OCR_receipts 병합) ==========

# Environment
python-dotenv>=1.0.0

# Web / API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
wsproto>=1.2

# Flask (recommendation)
flask>=3.0.0
flask-cors>=4.0.0

# Core Math & Data (numpy<2 for ChromaDB/Paddle 호환)
numpy>=1.26.3,<2
scikit-learn>=1.4.2
pandas>=2.2.0
joblib>=1.3.2

# Image & OCR (공통)
Pillow>=10.2.0
opencv-python-headless>=4.6.0.66
pytesseract==0.3.10
easyocr==1.7.1

# PaddleOCR
paddleocr>=2.7.0.3
paddlepaddle>=2.6.2

# OCR_receipts 전용 (이미지 전처리·ONNX)
imgaug
shapely
pyclipper
scikit-image
paddle2onnx>=1.0.6,<1.1
onnxruntime>=1.14.0

# ML (OCR_receipts)
xgboost>=2.0.3
lightgbm>=4.3.0

# 시각화 (선택)
matplotlib>=3.8.4
seaborn>=0.13.2

# LLM & Embedding
google-generativeai>=0.3.2
google-genai>=1.0.0
httpx>=0.26.0

# Schema & API
pydantic>=2.5.3

# DB & Vector
psycopg2-binary>=2.9.9
pgvector>=0.2.4
chromadb>=0.4.0

# LangChain
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
